I feel the need.  The need for... SPEED!

+++ tests/test_datamodule.py
@@ -5,6 +5,9 @@ import buffer_h5 as b5
 import os
 import numpy as np
 from env.debug import DummyEnv
+from datamodule import H5NextStateReward, write_balanced_splits, H5ImageLabelDataset, PolicyActionLabels
+from torch.utils.data import DataLoader
+from torchvision.transforms import ToTensor
 
 def identity(x):
     return x
@@ -25,7 +28,17 @@ def filename():
     if os.path.exists('test.h5'):
         os.remove('test.h5')
     yield 'test.h5'
-    os.remove('test.h5')
+    if os.path.exists('test.h5'):
+        os.remove('test.h5')
+
+
+@pytest.fixture
+def dest_filename():
+    if os.path.exists('load.h5'):
+        os.remove('load.h5')
+    yield 'load.h5'
+    if os.path.exists('load.h5'):
+        os.remove('load.h5')
 
 
 shape = (210, 160, 3)
@@ -57,9 +70,10 @@ def populated_buffer(filename):
 
     env = DummyEnv(traj)
     b = b5.Buffer()
-    state_col = b5.Column('state', (1, ), np.uint8, compression='gzip')
-    raw_col = b5.Column('raw', (240, 160, 3), np.uint8, compression='gzip')
-    action_col = b5.Column('action', dtype=np.int64, chunk_size=100000)
+    state_col = b5.NumpyColumn('state', (1,), np.uint8, compression='gzip')
+    raw_col = b5.NumpyColumn('raw', (240, 160, 3), np.uint8, compression='gzip')
+    grad_col = b5.NumpyColumn('grad', (240, 160, 3), np.uint8, compression='gzip')
+    action_col = b5.NumpyColumn('action', dtype=np.int64, chunk_size=100000)
     b.create(filename, state_col=state_col, action_col=action_col, raw_col=raw_col)
 
     def policy(state):
@@ -69,6 +83,8 @@ def populated_buffer(filename):
         if step + 1 == len(t1) -1 + len(t2) - 1 + len(t3) - 1 + len(t4)-1 + len(t5) - 1:
             break
 
+    grad_col.create(b.replay)
+    b.f['/replay/grad'].resize(b.steps, axis=0)
     return b
 
 
@@ -121,3 +137,39 @@ def test_r_distance(filename):
         x.append(b[item][1])
     assert count == 8
     b.close()
+
+
+def test_write_new(filename, dest_filename):
+    b = populated_buffer(filename)
+    b.close()
+    ds = H5NextStateReward()
+    ds.transforms = identity
+    ds.load(filename)
+    class_dict = {'no_reward': 0, 'reward': 1}
+    write_balanced_splits(ds, dest_filename, class_dict, 256, 10, 'gzip', 6)
+    batch_ds = H5ImageLabelDataset(dest_filename, 'train', ToTensor(), batch_size=256)
+    dl = DataLoader(batch_ds, num_workers=0, batch_size=None)
+    for image, label in dl:
+        assert image.shape[0] == 256
+        assert image.shape[1] == 6
+        assert image.shape[2] == 240
+        assert image.shape[3] == 160
+        assert image.shape[0] == 256
+
+
+def test_policy_action_labels(filename, dest_filename):
+    b = populated_buffer(filename)
+    b.close()
+    ds = PolicyActionLabels()
+    ds.transforms = identity
+    ds.load(filename)
+    class_dict = dict(zip(ds.name_classes, range(len(ds.name_classes))))
+    write_balanced_splits(ds, dest_filename, class_dict, 256, 10, 'gzip', 6)
+    batch_ds = H5ImageLabelDataset(dest_filename, 'train', ToTensor(), batch_size=256)
+    dl = DataLoader(batch_ds, num_workers=0, batch_size=None)
+    for image, label in dl:
+        assert image.shape[0] == 256
+        assert image.shape[1] == 6
+        assert image.shape[2] == 240
+        assert image.shape[3] == 160
+        assert image.shape[0] == 256
